# pro/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mimux API",
    "version": "v1",
    "summary": "The REST API of the mimux pro layer — mail read, written and moved by a machine.",
...
    "termsOfService": "https://mimux.dev",
    "contact": {
      "name": "mimux",
      "url": "https://mimux.dev",
      "email": "mattmezza@gmail.com"
    },
    "license": {
      "name": "Elastic Licence 2.0 (the pro layer this API belongs to)",
      "url": "https://github.com/mattmezza/mimux/blob/main/pro/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "{your-mimux-host}/api",
      "description": "Your own mimux install. mimux is self-hosted: there is no shared or central API host, so this URL is a template you fill in with the origin you serve mimux from.",
      "variables": {
        "your-mimux-host": {
          "default": "https://mail.example.com",
          "description": "Scheme and host of your mimux instance, without a trailing slash — e.g. https://mail.example.com, or http://localhost:8080 in development. The API is always mounted under /api on the same origin as the web client."
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "tags": ["Accounts"],
        "operationId": "listAccounts",
        "summary": "List accounts",
        "description": "Every configured account, with its live sync state and stored-mail counters.\n\nNot paginated: `next_cursor` is always `\"\"`.\n\n**Scope: `accounts:read`.**",
        "x-required-scope": "accounts:read",
        "responses": {
          "200": {
            "description": "The configured accounts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "next_cursor"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Account" }
                    },
                    "next_cursor": { "$ref": "#/components/schemas/NextCursor" }
                  }
                },
                "example": {
                  "data": [
                    {
                      "name": "personal",
                      "email": "me@example.com",
                      "state": "ok",
                      "last_sync": "2026-08-17T09:31:04Z",
                      "messages": 41208,
                      "unread": 12,
                      "folders": 17
                    }
                  ],
                  "next_cursor": ""
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/messages/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/MessageID" }],
      "get": {
        "tags": ["Messages"],
        "operationId": "getMessage",
        "summary": "Get one message",
        "description": "Full metadata plus the body forms you ask for, plus the attachment manifest.\n\nThe body is fetched from IMAP (through the client's body cache). If that fails — the account is offline, the message is gone from the server — the request still succeeds: `body` is omitted and `body_error` explains. That is deliberate, so an agent can always read metadata even while sync is broken.\n\nThe `html` form is the same sanitized document the reading pane renders, with one difference: external images are left in place, because the caller is a machine reading JSON rather than a browser that would leak a read receipt. Active content is still stripped.\n\n`attachments` is always present, `[]` when the message has none — or when listing them failed, which is not reported separately.\n\n`thread_id` and `thread_size` are not set on this endpoint.\n\n**Scope: `mail:read`.**",
        "x-required-scope": "mail:read",
        "parameters": [
          {
            "name": "body",
            "in": "query",
            "description": "Which body forms to include.",
            "schema": { "type": "string", "enum": ["text", "html", "both"], "default": "text" }
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "patch": {
        "tags": ["Messages"],
        "operationId": "patchMessage",
        "summary": "Change read, star, labels or folder",
        "description": "Applies every field present in the body, in order: read, starred, labels added, labels removed, folder move. Omitted fields are left alone.\n\nWrites land in the local store immediately and the IMAP side effect follows in the background — the same optimistic pattern as the web client, minus its 11-second undo window: an API caller that asks for a move means it, so the server-side move fires at once. A background IMAP failure is logged, not returned; flag writes are also retried by the sync loop's dirty sweep.\n\n`folder_id` must belong to the same account as the message; cross-account moves are a 400. Moving to the folder the message is already in is a no-op.\n\nThe response is the message re-read from the store after all changes, without `thread_id`/`thread_size`.\n\n**Scope: `mail:modify`.**",
        "x-required-scope": "mail:modify",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessagePatch" },
              "examples": {
                "markRead": { "summary": "Mark read and star", "value": { "read": true, "starred": true } },
                "label": { "summary": "Retag", "value": { "labels_add": ["todo"], "labels_remove": ["inbox-zero"] } },
                "move": { "summary": "Move to a folder", "value": { "folder_id": 8 } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The message after the change.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Messages"],
        "operationId": "deleteMessage",
        "summary": "Move to Trash",
        "description": "Moves the message to the account's `trash` special-use folder. **Nothing is erased** — this is the client's Delete, which is a move. There is no API for permanent deletion.\n\nStore first, IMAP move in the background, no undo window. An account without a trash folder is a 400 and the message stays put.\n\n**Scope: `mail:modify`.**",
        "x-required-scope": "mail:modify",
        "responses": {
          "200": {
            "description": "Moved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MoveResult" },
                "example": { "status": "ok", "folder_id": 4 }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/messages/{id}/archive": {
      "parameters": [{ "$ref": "#/components/parameters/MessageID" }],
      "post": {
        "tags": ["Messages"],
        "operationId": "archiveMessage",
        "summary": "Move to Archive",
        "description": "Moves the message to the account's `archive` special-use folder. Store first, IMAP move in the background, no undo window. An account without an archive folder is a 400 and the message stays put.\n\nTakes no request body.\n\n**Scope: `mail:modify`.**",
        "x-required-scope": "mail:modify",
        "responses": {
          "200": {
            "description": "Moved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MoveResult" },
                "example": { "status": "ok", "folder_id": 3 }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/messages/{id}/spam": {
      "parameters": [{ "$ref": "#/components/parameters/MessageID" }],
      "post": {
        "tags": ["Messages"],
        "operationId": "spamMessage",
        "summary": "Move to Spam",
        "description": "Moves the message to the account's `spam` special-use folder. Store first, IMAP move in the background, no undo window. An account without a spam folder is a 400 and the message stays put.\n\nTakes no request body.\n\n**Scope: `mail:modify`.**",
        "x-required-scope": "mail:modify",
        "responses": {
          "200": {
            "description": "Moved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MoveResult" },
                "example": { "status": "ok", "folder_id": 5 }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/v1/drafts": {
      "post": {
        "tags": ["Drafts"],
        "operationId": "createDraft",
        "summary": "Create a draft",
        "description": "Stores a draft — the same row the compose form autosaves into, so a draft created here opens in the web client.\n\nNote the shape difference against `POST /v1/messages/send`: recipients here are **single comma-separated strings**, not arrays, because that is how the compose form stores them.\n\n`in_reply_to` takes a mimux message id and is resolved to that message's `Message-ID` header on the way in; it also fills `account` when you left it blank. `kind` defaults to `new`.\n\nHonest caveat: there is no `GET /v1/drafts` or `DELETE /v1/drafts/{id}` yet. You can create and update; listing and deleting are web-client-only.\n\n**Scope: `mail:send`** — a draft is unsent mail, so it follows the sending scope rather than `mail:modify`.",
        "x-required-scope": "mail:send",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DraftCreate" },
              "example": {
                "to": "bob@example.com, carol@example.com",
                "subject": "Draft subject",
                "body": "Half a thought.",
                "mode": "plain"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "headers": { "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Draft" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "`in_reply_to` names no message.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/filters": {
      "get": {
        "tags": ["Filters"],
        "operationId": "listFilters",
        "summary": "List filter rules",
        "description": "Every stored rule, in no guaranteed order beyond what the store returns (rules carry a `position` that decides evaluation order).\n\nNot paginated: `next_cursor` is always `\"\"`.\n\n**Scope: `mail:read`** — filters govern how mail is handled, so they follow the mail scopes rather than getting one of their own.",
        "x-required-scope": "mail:read",
        "responses": {
          "200": {
            "description": "The rules.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "next_cursor"],
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Filter" } },
                    "next_cursor": { "$ref": "#/components/schemas/NextCursor" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Filters"],
        "operationId": "createFilter",
        "summary": "Create a filter rule",
        "description": "Validates and stores a rule. Any `id` in the body is ignored and overwritten by the assigned one.\n\nValidation is strict and happens before the write: a name is required, at least one condition and at least one action are required, condition fields and operators must be known, values must be non-empty, and a `regex` condition must actually compile. Failures come back as `400 invalid_request` with the validator's own message.\n\nRules apply to incoming mail during sync. Creating one does not re-run it over existing mail.\n\nHonest caveat: there is no update or delete endpoint for filters yet — creating and listing is the whole surface.\n\n**Scope: `mail:modify`.**",
        "x-required-scope": "mail:modify",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Filter" },
              "example": {
                "name": "Newsletters",
                "enabled": true,
                "conditions": [{ "field": "from", "op": "contains", "value": "news@" }],
                "actions": [{ "type": "move", "arg": "Newsletters" }, { "type": "mark_read" }]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, with its assigned id.",
            "headers": { "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Filter" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A mimux personal access token, minted in **Settings → API** and shown exactly once at creation. It always starts with `mimux_pat_`; a bearer value without that prefix is rejected without so much as a hash comparison.\n\n```\nAuthorization: Bearer mimux_pat_XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\n```\n\nTokens are stored as argon2id hashes, may carry an expiry, and can be revoked from Settings — revocation is honoured on the very next request, because the middleware re-reads the token row every time.\n\n**Scopes** are a space-separated subset of:\n\n| Scope | Grants |\n|---|---|\n| `mail:read` | Read mail: folders, messages, bodies, attachments, search, filters list. |\n| `mail:send` | Send mail and manage drafts. |\n| `mail:modify` | Change mail: read/star flags, labels, moves, archive/spam/trash, create filters. |\n| `accounts:read` | Read the account list. |\n| `webhooks:manage` | Manage webhook endpoints and their delivery log, including replaying a delivery and sending a test one. Reading them needs it too: a list of endpoints is a list of places this mailbox already talks to. |\n\nA token created with no recognised scope falls back to `mail:read`. Call `GET /v1/tokens/self` to find out what the token in your hand actually holds."
      }
    },
...
    "schemas": {
...
      "NextCursor": {
        "type": "string",
        "description": "Opaque cursor for the next page, or `\"\"` when the collection is exhausted. Endpoints that are not paginated always return `\"\"`."
      },
...
      "Account": {
        "type": "object",
        "description": "A configured account. Credentials — passwords, OAuth secrets, server hostnames — are never included.",
        "required": ["name", "email", "messages", "unread", "folders"],
        "properties": {
          "name": { "type": "string", "description": "The account's name, used wherever an `account` parameter is accepted." },
          "email": { "type": "string", "format": "email" },
          "state": {
            "type": "string",
            "description": "Live sync state. Absent when the account has no running worker.",
            "enum": ["ok", "syncing", "error"]
          },
          "message": {
            "type": "string",
            "description": "Detail accompanying `state`, typically the error when `state` is `error`. Absent when empty."
          },
          "last_sync": {
            "type": "string",
            "format": "date-time",
            "description": "Last *successful* sync, UTC. Absent until the account has synced once."
          },
          "messages": { "type": "integer", "format": "int64", "description": "Messages stored locally." },
          "unread": { "type": "integer", "format": "int64" },
          "folders": { "type": "integer", "format": "int64" }
        }
      },
...
    }
  }
}